texttransform: port tt bits 0x04 (HTML) and 0x01 (CR/CRLF), and fix UB the newly-reached paths exposed
Both remaining text-transform bits are now native, closing the tt gap for
-cc/-co/-cO.

0x01 -- TransformText_CR_to_CRLF (reference NZ_TextTransforms.cpp:402), applied
LAST in the chain, after 0x20/0x40. A small state machine that re-expands a
bare 10 into 13,10 (state 1) or rewrites a 10 back to a lone 13 (state 2). The
reference starts its output budget at out_cap + 1 and only notices the overrun
after writing that extra byte, so the port documents and the callers allocate
one byte of slack rather than silently overflowing; the returned size still
never exceeds out_cap.

0x04 -- HtmlTransformer (reference NZ_TextTransforms.cpp:781). The encoder
shortens a closing tag to a bare "</" and the decoder rebuilds the name from a
stack of opened tags, with "<//" as the escape for a literal "</". The stack's
lossiness is reproduced exactly (128 entries, names truncated to 16 bytes, a
4-entry recent-tag ring, and a fixed set of predefined tags that never get
stacked because they never need closing). Unlike the reference, tagnames_ and
taglens_ are zeroed so a malformed stream cannot read indeterminate stack
memory.

Also wires 0x20 into -cc, which the optimum path already had. That one is
explicitly NOT verified for -cc -- no -cc archive in the 60-file corpus sets
the bit, so there is no repro -- but it is the identical codec-agnostic
function (the reference applies 0x20 with no reorder_ascii_ dependency) and a
wrong result would decline on the entry checksum exactly as it does today. Both
tt gates now accept the same mask, removing an asymmetry that had no reason to
exist.

Getting these three files to decode pushed execution into tt16 and the CM
window matcher for the first time on this content, and ASAN/UBSAN immediately
flagged pre-existing UB there -- none of it in the new code:
  - nz_texttransform_num.cpp shifted a negative int left (twice). Now shifted
    as the unsigned bit pattern: bit-identical on two's complement, but no
    longer something the optimiser may act on. This is the same UB class that
    let gcc delete a loop bound in the BWT port.
  - nz_cm.cpp compared window bytes through unaligned uint16/uint32/uint64
    casts (6 sites). Now memcpy through small helpers -- one instruction on
    x86, defined everywhere.

Measured on one identical 60-file real corpus, baseline built at a3367ae:
-co 52 -> 55, -cO 54 -> 57, -cc 51 -> 54 (+3 each). Byte-exact on 9/9 repro
combinations (3 files x 3 codecs). native_only_v2.sh unchanged at 76/80 (no
synthetic fixture sets either bit). Suite 2553/2553, test_optimum_lz 4/4,
test_optimum2_lz 2/2, ASAN+UBSAN now clean on all 9 combinations, zero
regression, verified byte-exact against the binary built from this commit.

+++ src/nz_texttransform_num.cpp
@@ -271,7 +271,11 @@ struct TextTransformNumber {
             *b_ptr = (u16)(*b_ptr + ((is_predicted * 0x10006 - *b_ptr) >> 3));
             *c_ptr = (u16)(*c_ptr + ((is_predicted * 0x10000 - *c_ptr + 32) >> 6));
             one_.model_d_[d_idx >> 1] = (u8)(one_.model_d_[d_idx >> 1]
-                - (i32)(-(i32)(((u32)0x10 * is_predicted - d_value) >> 1) << (4 * (d_idx & 1))));
+                // Shift performed on the unsigned bit pattern: shifting a
+                // negative int is UB, and the optimiser is entitled to act on
+                // it (the same class of UB that let gcc delete a loop bound in
+                // the BWT port). Bit-identical on two's complement.
+                - (i32)((u32)(-(i32)(((u32)0x10 * is_predicted - d_value) >> 1)) << (4 * (d_idx & 1))));
         }
         if (!is_predicted) {
             if (num_zeros >= 4) {
@@ -317,7 +321,8 @@ struct TextTransformNumber {
                         *model_ptr = (u16)(*model_ptr + ((0x10000u * latest_bit - *model_ptr + 32u) >> 6));
                         u32 model_j_idx = model_j_upper + number_low;
                         one_.model_j_[model_j_idx >> 1] = (u8)(one_.model_j_[model_j_idx >> 1]
-                            - (i32)(-(i32)((22u * latest_bit - model_j_value) >> 3) << (4 * (model_j_idx & 1))));
+                            // Unsigned shift, see the matching note above.
+                            - (i32)((u32)(-(i32)((22u * latest_bit - model_j_value) >> 3)) << (4 * (model_j_idx & 1))));
                         number_low = number_low * 2 + latest_bit;
                     }
                     if (!latest_bit) lower = number; else upper = number;
